arm64: dts: ls1088ardb: describe the on-board TI retimer
commit  e3179dc772d9bf2f3835186aea5618d2119ed92c from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Now that we have a driver for the on-board TI retimer, we can fully
describe it in DT and take advantage of it by referencing it from the
ethernet node.

With this patch in place, we are now able to dynamically reconfigure
(only at boot time, since no MOD_ABS GPIO connected to the cage) the
DPMAC.1 ethernet interface depending on the SFP module inserted.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>