EDAC/versal: Fix the error injection
commit 71bc25031bfa3ad358bced47e14219e2ea242480 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Adjustments are made to the error injection bits configuration in the Versal
EDAC driver. The correction ensures that the error injection operates accurately
within the specified ranges. Previously, there was an oversight in the bit
position calculations, which has been rectified in this patch.
The flip0 writes 0 to 31 and flip1 has bits for 32 to 63.

Fixes: 14c3e181438f ("EDAC/versal: Make the bit position of injected errors configurable")
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>