HSD #15015532380: fpga: altera-cvp: Fixed credit register
commit 5b38b4adad9946a2af18f81e15f1bee069408975 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

This fix is required in order to get  Agilex5 SOC FPGA CVP bitstream
transfer done with the new credits changes.
Fixed the Agilex5 credits to be 12 bits instead of 8 bits in non-Agilex5
devices. Also added debug messages to make it clear to the user
that the currently attached device type family as well as status
of the CVP completion.

Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>