thermal: qoriq: Implement workwaround for TKT0635774
commit 2bff62b696f7131f1b2fa4ee3b9306c8249a32c5 from
https://github.com/nxp-auto-linux/linux

TMU is reporting very randomly sometimes large errors in temperature
measurement, for example, the temperature will go up by 40 degrees on
one sample and then down 40 degrees on the next sample.

One software workaround is based on using Monitor Rising and Falling
Temperature Rate Capture registers to filter out these kind of erroneous
measurements:
- Configure Monitor Rising/Falling Temperature Rate Critical Threshold
with a value configured in the DTS.
- When an interrupt occurs, check first if any critial rate threshold was
exceed. If yes, wait for next sampling to occur.
- Check the opposite rate register. If the difference between the rising
and falling rate is less than 2, this is considered an erroneous
measurement. The value of 2 was chosed based on the fact that the
difference between two consecutive samples should not be greater than 2
while in continuous monitoring. The workaround proposal is to consider a
glitch when the difference between rising and falling is less than 10%
but this works if the rate threshold is > 10.

ALB-10916
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>