fs/restrl: Allow the overflow handler to be disabled
commit d0fda356c1a32888dd939a451dd5aecada094fac from
git@git.assembla.com:cavium/WindRiver.linux.git

Resctrl has an overflow handler that runs on each domain every second
to ensure that any overflow of the hardware counter is accounted for.

MPAM can have counters as large as 63 bits, in which case there is no
need to check for overflow.

Call the new arch helpers to determine this.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>