drivers:mmc:cadence: Calculate delay-element
commit 217cf540a9a842761b7ea49f73da5d1d61e1b5f4 from
git@git.assembla.com:cavium/WindRiver.linux.git

The Cadence SDHCI driver uses a delay element count when the DLL
does not lock.  This patch allows the driver to automatically
calculate the number of delay elements to use to meet the timing
rather than use hard-coded values.  This allows for variation between
chips.  The delay element is not used when the DLL is locked.

Change-Id: I7d8c27612831bf981709ce702061f7924fe5cae7
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/124215
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>