staging: fsl_qbman: implement timeout on waiting for FQRN in qm_shutdown_fq()
commit  97f51799b999be7ce936443de75c6f1baeb07db8 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

It can happen for the code to hang waiting for a frame queue retire
notification, mainly because the FQ is enqueued onto a channel's WQ,
channel which is serviced through a software portal other than the one
we're polling through.

It is obnoxious when the driver hangs like that, because we call
qm_shutdown_fq() early on, from __init qman_init(), and we won't see a
thing unless the boot parameters contain "earlycon". So rather than
hang the kernel there, print something out and let the code execution
limp on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>