drivers/net: dwmac-s32cc: Add `clk_set_rate()` for GMAC0_RX_RGMII
commit 445d07f0ee5682db96fc71ff5b0e5386af9fa24f from
https://github.com/nxp-auto-linux/linux

Due to a different implementation of the clock driver in the M7 SRM
component, it requires that each clock that needs to be enabled/
have a rate set should have a specific clock framework call issued
from the A53 side.

Thus, added a `clk_set_rate()` call for the RX RGMII clock before
enabling it.

Also changed a print on the error path to `dev_err()`.

Issue: ALB-11086
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>