video/hdmi: Add support for version 3 AVI Infoframe
commit 88cd8fb8e29d37be9070a0fa57b5a72f7bdae6d5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

With the current version 2 AVI Infoframe support, when source sends
resolutions above the VIC>127, sink reads it as wrong. This is
because bit 7 of data byte 1 in version 2 AVI Infoframe is set to '0'.
Where as in version 3 AVI Infoframe is it set to '1'. Fix the video_code
mask to read all the 8bits for data byte 1.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>