iio: adc: s32cc_adc: fix syncing of dma region
commit 330cf786e2e16b7ad8c5a67af9f6b3025f42d2de from
https://github.com/nxp-auto-linux/linux

When using CONFIG_DMA_API_DEBUG, the kernel throws warning about the
ownership of DMA memory.
This happens because the syncing of the dma region is done by the ADC
driver. This is incorect because that region is allocated to the dma
driver.

Use the dma device instead of the ADC device.

Issue: ALB-11315
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>