vc4/hdmi: Ignore hotplug interrupt with force_hotplug
commit  451bfe943a19e368fdf28fba0be8641fdff7c7ed from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The intention of the vc4.force_hotplug setting is to
ignore hotplug completely.

It can be used when a display toggles hotplug when
switching AV inputs, going into standby or changing a
KVM switch, and some side effect of that is unwanted.

It turns out while vc4.force_hotplug currently makes
hotplug always read as asserted, that isn't enough to
stop drm doing lots of stuff, including re-reading
the edid.

An example of what drm does with a hotplug deasert/assert
and vc4.force_hotplug=1 currently is:

https://paste.debian.net/hidden/dc07434b/

That is unwanted. Lets ignore the hotplug interrupt
completely so drm is blissfully unaware of the hotplug change.

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>