LF-11973: arm64: dts: imx95-15x15-evk: add FTS touch support
commit  93935124f0e003563c985e6d82f6ecb1cdb58aa8 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Support FTS touch on the i.MX95 15x15 EVK platform.

There is a limitation for this touch on i.MX95 15x15 EVK
platform. This touch needs a reset signal to active itself.
However the display reset pin and touch reset pin are shared,
so touch driver needs to be probed after the display reset.

The safest method is the touch driver is installed as a module
instead of build-in on imx95-15x15-evk.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>