HSD #15015934876: dts: socfpga_agilex5.dtsi: Fix the clk-csr for XGMAC
commit 6b4a85dc4b3106c7ff7143b61b4483f84be6c532 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.51-lts.

This patch is to to set the clk-csr divider value. The value is
set based on the clk_csr_i clock speed and currently the hardware
is connected with the l4_sp_clk (100MHz) thus the clk-csr divider
need to choose the divider with the value 62, due to that, the
csr-clk value need to choose divider option 0.

Signed-off-by: Boon Khai Ng <boon.khai.ng@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>