gpio: s32: don't try to configure the GPIO when setting the IRQ type
commit f5aba477685277efaaf41a19ee5777588d80d216 from
https://github.com/nxp-auto-linux/linux

Trying to configure the GPIO when setting the IRQ type would lead to
an error. This is because a mutex from the pinctrl subsystem is
requested from a sleeping context.

There is no need to configure the GPIO as an input at that point, it
is already configured before the call to set the IRQ type.

Issue: ALB-11307

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>