spi: cadence-xspi: Fix locking with ATF
commit f25c023e9bf992bcde2560473e4fcc389ed8096f from
git@git.assembla.com:cavium/WindRiver.linux.git

The locking code with ATF is flawed where if both the kernel and ATF
try and grab the lock at the same time the kernel driver will not
detect this since it does not re-check the lock value.  This also
simplifies the code a bit by using a for loop instead of a while loop.

This lock is necessary since both ATF and the kernel can require access
to the SPI NOR and the GPIO register is otherwise unused so it is
shared between both.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/127790
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Change-Id: Ia87aed564e5d605c4bbabd94f91c7d353dac1d1f
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>