untested: arm_mpam: resctrl: Add support for csu counters
commit b47661356a0cecef0222d714a0fc56b5224daeb4 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl exposes a counter via a file named llc_occupancy. This isn't really
a counter as its value goes up and down, this is a snapshot of the cache
storage usage monitor.

Add some picking code to find a cache as close as possible to the L3 that
supports the CSU monitor.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>