x86/resctrl: Add end-marker to the resctrl_event_id enum
commit 7a2a6301b96a9a5973209f0e47754f2249c3dc10 from
git@git.assembla.com:cavium/WindRiver.linux.git

The resctrl_event_id enum gives names to the counter event numbers on x86.
These are used directly by resctrl.

To allow the MPAM driver to keep an array of these the size of the enum
needs to be known.

Add a 'num_events' define which can be used to size an array. This isn't
a member of the enum to avoid updating switch statements that would
otherwise be missing a case.

Signed-off-by: James Morse <james.morse@arm.com>
Tested-by: Carl Worth <carl@os.amperecomputing.com> # arm64
Tested-by: Shaopeng Tan <tan.shaopeng@jp.fujitsu.com>
Reviewed-by: Shaopeng Tan <tan.shaopeng@jp.fujitsu.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>