tty: serial: fsl: add separate write_atomic callback
commit 67744068a359293f365660731e6d993f3057fb2e from
https://github.com/nxp-auto-linux/linux

The previous write_callback acuqires a spinlock which on full PREEMPT_RT
is changed into an rt_mutex. Therefore, it results in an invalid
possible wait during an atomic context.

Add a separate callback which uses the console_lock_atomic function
provided for this case. Also, don't call uart_write_wakeup because it
also takes a spinlock. Most probably, it also isn't necessary to call it
for the write callback since it relates to console outputting but I will
keep it to preserve the original implementation.

Issue: ALB-11307

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>