backlight: Add a display name to the core, and a function to set it
commit  a9e2472e480b35e591d8511f76c50d81703d45e0 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The naming of backlight devices is not terribly useful for
associating a backlight controller with a display (assuming
it is attached to one).

Add a sysfs node that will return a display name that can be set
by other subsystems.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>