clocksource: Add option to force acpi_pm as clocksource watchdog
Add a kernel command line option to force the use of acpi_pm as the
watchdog clocksource. To use it, add the following to the
kernel command line:

     ricardo_clocksource_wd=force_acpi_pm_wd

and in dmesg look for:

     RICARDO: Force acpi_pm as watchdog

Also, add a debugfs entry to inspect the current clocksource watchdog
and the history watchdoug sources. The file is

/sys/kernel/debug/ricardo_clocksource/watchdog_info

It look like something like:

CURRENT: acpi_pm
HISTORY:
apci-pm

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
[YL: Original patch from
https://github.com/intel-collab/networking.wireless.transport.rdk.board-support.meta-intel-axxia.git.]
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>