octeontx2-pf: Configure VF mtu via representor
commit 9a0c3085f5d50b2b1ba95aad421f9cdc2a7d7787 from
git@git.assembla.com:cavium/WindRiver.linux.git

Adds support to manage the mtu configuration for VF through representor.
On update of representor mtu a mbox notification is send
to VF to update its mtu.

This feature is implemented based on the "Network Function Representors"
kernel documentation.
"
Setting an MTU on the representor should cause that same MTU
to be reported to the representee.
"

Signed-off-by: Sai Krishna <saikrishnag@marvell.com>
Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Reviewed-by: Simon Horman <horms@kernel.org>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>