drivers: mmc: core: handle card-removal when running CQE recovery
commit 62640cd5ffe0cc3986c24faa1f8b92d848bab4e4 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Recovery claims the MMC card so the card-detect work gets significantly
delayed - leading to lots of error recovery loops that can never do
anything but fail.

Explicitly detect the card after CQE has halted and bail if it's not
there.

Also ratelimit a not-very-descriptive warning - one occurrence in dmesg
is enough to signal that something is amiss.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>