LF-12679-2 media/i2c: mx95mbcam: Reset serializer before removal or reboot
commit  e6871656cda0b205a6ff70d5d6181446a623c39d from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Since the MX95MBCAM sensor module has the PWDNB pin, that should reset
the serializer, hard-wired to HIGH, we need to make sure we leave the
serializer in the default state when we remove the module or we perform
a warm reboot. This would avoid issues with GMSL links not locking when
rebooting the board.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>