arm_mpam: Quirk CMN-650's CSU NRDY behaviour
commit b970a9a458e309c0433df05aad06b3d33d121cd9 from
git@git.assembla.com:cavium/WindRiver.linux.git

CMN-650 is afflicted with an erratum where the CSU NRDY bit never clears.
This tells us the monitor never finishes scanning the cache. The erratum
document says to wait the maximum time, then ignore the field.

Add a flag to indicate whether this is the final attempt to read the
counter, and when this quirk is applied, ignore the NRDY field.

This means accesses to this counter will always retry, even if the
counter was previously programmed to the same values.

The counter value is not expected to be stable, it drifts up and down
with each allocation and eviction. The CSU register provides the value
for a point in time.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>