MMFMWK-9397: arm64: wave6: remove level interface
commit  98bb00fe3c7051074d3c34ad7ee79b41ae91c681 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

For 4k decoding in our board, the performance can reach higher level than
which marked in datasheet.
So removing level interface of decoder is reasonable.
Otherwise, We can not ensure that the encoder will not use a large level
when encoding with a small bitrate or resolution.

Signed-off-by: TaoJiang <tao.jiang_2@nxp.com>
Reviewed-by: MingQian <ming.qian@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>