gpio: siul2: populate num_reg_defaults_raw for output regmaps
commit 3d12e316a5c3886e65a576b3103b663315a1381b from
https://github.com/nxp-auto-linux/linux

Previously, the cache for the PGPDO registers would be populated with 0
because no reg_defaults/reg_defaults_raw were given to the
regmap_config.

Force the regcache to read the hardware state by passing in
num_reg_defaults_raw set to the number of registers for that regmap.

Issue: ALB-11725

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>