drm/microchip: add a driver for the Microchip GFX2D GPU
commit d70242a8c2468686412bce3de53ed7b622a4fd04 from
https://github.com/linux4microchip/linux.git

The Microchip GFX2D is a 2D GPU, which supports very basic graphics
instructions. Each intruction process a single rectangle defined by:
- its width and height.
- the coordinates of its top-left corner relative to each surface
  involved in the operation.

Indeed, each graphics instruction works on one surface at least, the
target surface where the rectangle is drawn. More surfaces may be
required as sources depending on the type of graphics instruction.

The supported graphics instructions are:
- FILL: uniformly fills the rectangle in the target surface with a
  color.
- COPY: copies the rectangle from one source surface into the target
  surface.
- BLEND: blends the rectangles from two source surfaces into a single
  rectangle within the target surface, according to blend parameters
 (blend function, blend factors, blend colors, ...).
- ROP: operates on rasters taking two or three surfaces as inputs with
  additional parameters and fills the rectangle in the target surface
  based on the requested raster operation.

GFX2D instructions are ordered by being queued into a ring buffer in
the system memory. This driver updates the ring buffer head pointer to
notify the GPU about new instructions to process, while the GPU updates
the tail pointer when the instructions have been fetched (but not
necessarily completed yet).

A trick with LDR (load) and STR (store) instructions has been
implemented to precisely know whether a given GFX2D instruction has
completed. Indeed, with its interrupt or status register only, the GFX2D
can only notify that some instruction has completed but that's not
enough to find out which one precisely.

Therefore, after each batch a graphics instructions named as a command,
a LDR instruction loads the command ID (a uint32_t monotonically
increasing) into a GFX2D internal general purpose register.
This LDR instruction is immediately followed by a STR instruction, which
writes the value in the GFX2D internal register, hence the command ID,
into some fixed address in the system memory which is used as a mailbox.
Also, when the STR instruction completes, it triggers an interrupt in the
GFX2D, if this interrupt is enabled, to wake up the interrupt thread as
needed and to serve waiting tasks.

So, at any moment, the driver can read the mailbox; any command with an
ID lower than or equal to the value read in the mailbox is guaranteed to
be completed by the GPU. Actually, integer overflows are handled by
comparing both the mailbox value and the command ID to a third value,
the ID of the latest command completed by the driver, which is known to
be lower than or equal to both the mailbox value and the ID of the
command being tested for completion.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>