arm_mpam: resctrl: Add support for 'MB' resource
commit 89c44d93341f72bd9567accb1a5116f942d7cb34 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl supports 'MB', as a percentage throttling of traffic somewhere
after the L3. This is the control that mba_sc uses, so ideally the
class chosen should be as close as possible to the counters used for
mba_local.

MB's percentage control can be backed either with the fixed point
fraction MBW_MAX or the bandwidth portion bitmap. Add helper to convert
to/from percentages. One problem here is the value written is not the
same as the value read back. This is deliberatly made visible to
user-space. Another is the MBW_MAX fixed point fraction can't represent
100%. This is going to confuse user-space, so shift everything up taking
all-ones as 100% and zero as the minimum granularity.

Signed-off-by: James Morse <james.morse@arm.com>>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>