dpaa2-eth: extend CEETM support to 16 Tx TCs
commit  d8d0d0d8b1976d05f6e4cf4c0e819776f3db7cd0 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The HW available on the LX2160A SoC supports 16 Tx TCs. This patch
extends the CEETM qdisc support to cover this usecase.

First of all, all the macros which hold the maximum number of Rx or Tx
TCs are updated to reflect the new behavior. Since starting with this
patch the number of Rx and Tx TCs could be different from one another,
the call to alloc_etherdev_mq() is replaced by alloc_etherdev_mqs() so
that we can separately can configure the two types of queues.

In case the underlying DPNI object has a number of Tx TCs greater than 8
(configurable at dpni_create() time), the CEETM qdisc will impose the
following restrictions:
- The first 8 TCs can only be configured in strict priority mode.
- TCs [8-12) can only be part of WEIGHTED_A group
- TCs [12-16) can only be part of WEIGHTED_B group

The behavior of the underlying dpni_set_tx_priorities() MC API changes
in case the num_tx_tcs > 8 in the following way: the cfg.tc_sched[]
array will hold the configuration requested for the TCs >= 8 since the
first 8 TCs are in a strict priority which cannot be changed.
This means that the CEETM qdisc behavior was changed so that it
configures the appropriate tc_sched[] structure either the one found at
index 'qpri' or at index 'qpri % 8', depending on the number of Tx TCs.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>