net: sdk_dpaa: disable multiple technology abilities for backplane
commit  9951ce71d531fbbf356db5e22eb7fe2f2e6a819d from
https://github.com/nxp-imx/linux-imx lf-6.18.y

We are not prepared to handle this in sdk_fman's adjust_link() methods,
and this is now the only driver incapable of doing so.

Let the mtip_backplane continue to probe for all backplane modes, but
limit the supported mask at the MAC side.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>