gpio: scmi: don't dispose irq_mappings
commit 262e99adda9d1cf15873d988efdaec6dbbfb66e7 from
https://github.com/nxp-auto-linux/linux

Irq mappings are created in the .to_irq callback. They don't need to be
removed in the `scmi_gpio_free`. They are disposed of when the
gpiochip is removed. Moreover, if an interrupt is freed and
re-requested, it will first try to find if an available irq_mapping
exists and use it if it exists or create a new one otherwise.

Issue: ALB-11575

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>