drm/vc4: Add a delay after disabling hdmi phy output
commit  126de0d8e268a696c6245cfbbedac827548cdef4 from
https://github.com/raspberrypi/linux.git rpi-6.18.y

There appears to be a requirement for some devices
(I'm testing with a 8K VRROOM 40Gbps HDMI switch)
for a measable delay between removing the hdmi phy output from
the old mode, to enabling the hdmi phy output for the new mode.

Without the delay, a mode switch has a small change of getting a permanent
'no signal', which requires a subsequent mode switch or a unplug/replug
to redetect.

Switching between 4kp24/25/30 modes fails about 5% of time in my testing.

Add a delay to make it impossible to switch faster than this.

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>