arm64: zynqmp: Expose AMS to userspace as HWMON
commit 55858c7d2e7a6b39ad82dfa861e55fe6a9c239f0 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Expose the AMS to userspace, allowing monitoring of internal voltages
and temperatures. For compatibility, we keep the node name the same as
on the SM-K26, and we keep the ZCU100 Rev C. around (since it is named
differently).

Signed-off-by: Sean Anderson <sean.anderson@linux.dev>
Link: https://lore.kernel.org/r/20240812215129.3599832-3-sean.anderson@linux.dev
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: upstream (e279e4b2faf81eb5570138eb61f2c8acd95ce893)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>