arm64: dts: fsl-lx2160a: include rev2 chip's dts
commit  85532dc6ad689418307eaa1acb7241b99d337581 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The mass production LX2160A silicon is rev2, and it uses the DesignWare
PCIe Controller. Old Rev1, which used the Mobiveil PCIe controller, was
not supported.

Although the U-Boot fixup can change compatible string fsl,lx2160a-pcie
to fsl,ls2088a-pcie since 2019, it is confusing for the in-tree device
trees to be only for Rev 1 silicon.

Change the NXP development boards to include rev2's dtsi by default.
If ever necessary, "-rev1" variants of device trees can be added back
for the retired SoC.

We don't know how customer boards are dealing with this situation, so
leave it up to them to decide, and make no changes there.

Link: https://lore.kernel.org/lkml/20240826-2160r2-v1-3-106340d538d6@nxp.com/
Signed-off-by: Frank Li <Frank.Li@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>