hwmon: (ina2xx) Add support for INA260
commit cf2fb7a72d35637c98a97f4ba78721415f1a7398 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

INA260 is similar to other chips of the series, except it has an internal
shunt resistor. The calibration register is therefore not present. Also,
the current register address was changed, though that does not matter for
the driver since the shunt voltage register (which is now the current
register) value is already used to read the current.

Cc: Loic Guegan <loic.guegan@mailbox.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@kernel.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
State: upstream (70fb84a109c639637f0636281dbdb21ed8ffb000)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>