gpio: scmi: remove irq_domain ops
commit 82751e65d604c315b66a73c10f4d42ee6f4689a0 from
https://github.com/nxp-auto-linux/linux

The default gpiochip implementations were used apart from the xlate
callback which also tried to update the mapping for an interrupt.

However, the documentation for the xlate callback specifies that it
should only be used for decoding the hardware irq number and the
interrupt type. This can be achieved by calling the gpiochip default
implementation.

Issue: ALB-11799

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>