octeontx2-af: cn20k: Add new mailboxes for CN20K silicon
commit 5c347f92e962caa9a7955921d29db64023213bf6 from
git@git.assembla.com:cavium/WindRiver.linux.git

Due to new requirements for CN20K, the existing structure "struct
mcam_entry" needed to be updated. Earlier it was using two arrays "kw"
and "kw_mask" of length 7(key word size). To accommodate CN20K
requirements, the size of the arrays are updated from 7 to 8. But, with
this change, the backward compatibility is broken because the size of the
structure is changed in between. Hence, we decided to use separate
mailboxes which involves the structure "struct mcam_entry". This patch
identifies such mailboxes and adds new mail boxes for CN20K specific
requirement.

Four new mailboxes:
1. NPC_CN20K_MCAM_WRITE_ENTRY
2. NPC_CN20K_MCAM_ALLOC_AND_WRITE_ENTRY
3. NPC_CN20K_MCAM_READ_ENTRY
4. NPC_CN20K_MCAM_READ_BASE_RULE

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: Idd0b90e16c0e13d97df62852ddbb2c06cf55c7ba
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/134632
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>