LF-14260-4 net: enetc: add Egress Sequence Recovery Table support
commit  34e32ae09b0c5aeb0eb866269378120f7f562ea5 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The Egress Sequence Recovery table contains one entry per FRER recovery
function (implemented as per 802.1CB), and each entry contains a number
of configuration elements, whose value is set when the table entries are
configured, and operational elements, which are managed by the hardware
and can be queried by software. Whether or not a FRER recovery function
is to be performed is specified in the Egress Treatment table, and if so
then the Egress Treatment table also provides an index into a Egress
Sequence Recovery table.
The Egress Sequence Recovery table is implemented as a linear array of
entries accessed using an index (0,1, 2, ..., n) that uniquely identifies
an entry within the array.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>