LF-14260-5 net: enetc: add Egress Count Table support
commit  84b3b7848599cdad9ae1c6be1ff8038f846ede31 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Egress related statistics are maintained in this table. The Egress
Count table is implemented as a linear array of entries accessed
using an index (0, 1, 2, ..., n) that uniquely identifies an entry
within the array. Egress Counter Entry ID (EC_EID) is used as an
index to an entry in this table. The EC_EID is specified in the
Egress Treatment table. Egress Count table entries are always
present and enabled.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>