octeontx2-af: Fix the issue of infinite loop in nix_aq_reset
commit 80c40f54bcfe2c3d2d2873b314e19a719e2a7fb1 from
git@git.assembla.com:cavium/WindRiver.linux.git

In nix_aq_reset, the while loop that checks busy bit in NIX_AF_AQ_STATUS
register is not breaking out of the loop if the busy bit is not cleared.
This results in infinite loop. Fix this issue by breaking out of the loop
if the busy bit is not cleared & timeout is reached.

Fixes: a77b91137260 ("octeontx2-af: Dump hw register state on error")
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Change-Id: I555c92e3b96c101cca988c64a61566c3b38ef93e
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/135386
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>