iio: s32cc_adc: Add s32cc_adc_stop_conversion function
commit 47e2066e1e612973c21516e643362c9fac88820b from
https://github.com/nxp-auto-linux/linux

This function only stops the conversion of data.

After the NSTART bit is reset, the ADC will go to IDLE but we still
have to wait for the conversion to finish since we might have DMA
transactions in progress. If ADC still converts and we stop the hardware
DMA service requests, DMA might reach an inconsistent state and never
restart again.

Issue: ALB-11737
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>