LF-13746-4 arm64: dts: imx95: update the netc_blk_ctrl node
commit  4a26a5d25bfcea44996dac549280acffaa3f1832 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Since the compatible string of netc_blk_ctrl has been changed in
both nxp,imx95-blk-ctl.yaml and imx95-blk-ctl driver, so we need
to update the compatible string in the DTS to make the clk driver
works properly.
In addition, for the imx95-blk-ctl clock driver, only the first 2
registers in NETCMIX BLOCK CONTROL are used to configure the clock,
and the other parts are not related to the clock, so correct the
register address length.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>